re-organize documentation with new sections
