+++ website/documentation/sections/binding_properties.py
@@ -1,0 +1,5 @@
+from ..tools import load_demo
+
+
+def content() -> None:
+    load_demo('bindings')
